增加对float的处理

+++ Assets/Editor/CustomMap.cs
@@ -7,7 +7,7 @@ public class CustomMap : ScriptableObject
 {
     public int mapwidth;
     public int mapheight;
-    public int unitlength;
+    public float unitlength;
     public int paintedgridlength;
   //  public int tilelength;
     public Vector3 dir;
